+++ cmd_queue/__init__.py
@@ -306,7 +306,7 @@ Example:
 __mkinit__ = """
 mkinit -m cmd_queue
 """
-__version__ = '0.2.1'
+__version__ = '0.2.2'
 
 
 __submodules__ = {